requirements.txt
djangorestframework
Pillow
python-dotenv
django-rest-swagger